requirements.txt
python-telegram-bot==21.*
APScheduler==3.*
httpx
python-dotenv
